scraper_api/requirements.docker.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
playwright==1.58.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dotenv==1.2.2
python-telegram-bot==22.7
PyYAML==6.0.3
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
psycopg[binary]
slowapi==0.1.9
gunicorn==23.0.0
asyncpg==0.30.0
